- create show for motorcycles with all details nested in one page - remove CascadeType.ALL and orphanRemoval = true from the one to many relationships - refactor delete controllers of models with a one to many relationship

+++ backoffice/src/main/java/com/honda/webapp/backoffice/controller/MotoController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.honda.webapp.backoffice.model.Moto;
 import com.honda.webapp.backoffice.repository.CategoryRepository;
@@ -110,12 +111,20 @@ public class MotoController {
   }
 
   @PostMapping("/delete/{id}")
-  public String delete(@PathVariable Integer id) {
+  public String delete(@PathVariable Integer id, RedirectAttributes redirectAttributes) {
 
     Optional<Moto> motoAttempt = motoService.findById(id);
 
     if (motoAttempt.isEmpty()) {
       return "404";
+    }
+
+    Moto moto = motoAttempt.get();
+
+    if (!moto.getVariants().isEmpty()) {
+      redirectAttributes.addFlashAttribute("error",
+          "You can't delete a motorcycle which has variants associated with it");
+      return "redirect:/motos";
     }
 
     motoService.deleteById(id);
